refactor(jina): 관리자 페이지 page 디렉토리로 이동
